lambdas/wx_bootstrap/requirements.txt
requests==2.32.3
PyJWT==2.9.0
cryptography==43.0.1
boto3
